+++ src/assets/scss/page/_cart.scss
@@ -21,8 +21,65 @@
     }
 
     &-container {
+        display: flex;
+    }
+
+    &-list {
+        flex-basis: 65%;
+    }
+
+    &-item {
         display: grid;
-        grid-template-columns: 2fr 1fr;
+        grid-template-columns: 2fr 1fr 1fr;
+        place-items: center;
+        padding-top: 20px;
+    }
+
+    &-flavor {
+        display: flex;
+    }
+
+    &-img {
+        flex-basis: 30%;
+    }
+
+    &-goods {
+        flex-basis: 70%;
+    }
+
+    &-maker {
+        padding-top: 15px;
+        font-size: 14px;
+    }
+
+    &-name {
+        padding-top: 10px;
+        font-size: 18px;
+    }
+
+    &-price {
+        padding-top: 10px;
+        font-size: 16px;
+
+        span {
+            font-size: 12px;
+        }
+    }
+
+    &-total {
+        text-align: right;
+        font-size: 22px;
+        span {
+            display: inline-block;
+            padding-left: 10px;
+            font-size: 12px;
+        }
+    }
+
+    &-side {
+        flex-basis: 35%;
+        height: 300px;
+        border: 2px solid #000;
     }
 
 }
