+++ src/assets/scss/page/_top.scss
@@ -49,6 +49,8 @@
         display: flex;
         justify-content: space-between;
         align-items: center;
+        width: 120px;
+        height: 60px;
         border: 1px solid #000;
     }
